@@ -191,4 +191,4 @@ const WebVTTParser = {
     }
 };
 
-module.exports = WebVTTParser;
+export default WebVTTParser;
